% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenbasis.R
\name{eigenbasis}
\alias{eigenbasis}
\title{Eigenbasis selection}
\usage{
eigenbasis(X, method = "FVE", FVEthreshold = 0.9)
}
\arguments{
\item{X}{data matrix, functional values evaluated at common domain,
one function per row}

\item{method}{method used for the selection of the number of eigenvectors,
possible values are \code{FVE} for the fraction-of-variance-explained in the
fitted covariance function, or \code{BIC} for the conditional pseudo-Gaussian BIC selection}

\item{FVEthreshold}{a numerical constant between 0 and 1, giving the threshold
for method \code{FVE}. Ignored with \code{method=BIC}.}
}
\value{
A matrix whose columns correspond to the estimated basis of functions.
}
\description{
Automated eigenbasis selection for densely observed functional data.
}
\examples{
gridsize = 101                           # grid size
Grid = seq(0, 1, length = gridsize)      # grid of measurements
n = 150                                  # sample size
Jmodel = 4                               # number of basis elements used for building functional predictors
BASIS = phif(Jmodel, Grid)               # first Jmodel Fourier basis elements
UNIF = matrix(runif(n * Jmodel, min = -1, max = 1), nrow = n, ncol = Jmodel)
X = UNIF \%*\% BASIS[1:Jmodel, ]

# dimensions selected using different methods
dim(eigenbasis(X))
dim(eigenbasis(X,FVEthreshold = .95))
dim(eigenbasis(X,method="BIC"))
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
