% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{expand}
\alias{expand}
\title{Basis expansion}
\usage{
expand(C, basis, na.to.zero = TRUE)
}
\arguments{
\item{C}{a vector of coefficients, or a matrix with one vector per row}

\item{basis}{a matrix of basis functions, one function per row}
}
\value{
A matrix of functional values that corresponds to the coefficients given
in the rows of \code{C} with respect to the basis in \code{basis}.
}
\description{
Expand a set of coefficients into functions in a basis.
}
\examples{
gridsize = 101                           # grid size
Grid = seq(0, 1, length = gridsize)      # grid of measurements
n = 150
Jmodel = 4                               # number of basis elements
BASIS = phif(Jmodel, Grid)               # first Jmodel Fourier basis elements
UNIF = matrix(runif(n * Jmodel, min = -1, max = 1), nrow = n, ncol = Jmodel)

PREDICTORS = UNIF \%*\% BASIS[1:Jmodel, ]
PREDICTORS2 = expand(UNIF,BASIS[1:Jmodel, ])
all.equal(PREDICTORS,PREDICTORS2)        # expansion of coefficients into functions

}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
