% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAM.R
\name{FAM}
\alias{FAM}
\title{Functional additive model estimation}
\usage{
FAM(
  Responses,
  LEARN,
  PRED,
  method = "FVE",
  FVEthreshold = 0.9,
  BASIS = NULL,
  knn.ind = TRUE,
  fullgrid = FALSE
)
}
\arguments{
\item{Responses}{vector of scalar responses}

\item{LEARN}{matrix of regressors, functional values evaluated at common
domain, one function per row}

\item{PRED}{matrix of functions at which new values are predicted, one
function per row}

\item{method}{method used for the selection of the number of eigenvectors,
possible values are \code{FVE} for the fraction-of-variance-explained in the
fitted covariance function, or \code{BIC} for the conditional pseudo-Gaussian
BIC selection}

\item{FVEthreshold}{a numerical constant between 0 and 1, giving the
threshold for method \code{FVE}. Ignored with \code{method=BIC}.}

\item{BASIS}{if basis functions are not estimated using the empirical
covariance operator, matrix of basis functions. One function per row.}

\item{knn.ind}{logical; should the bandwidth be chosen using the k-NN
cross-validation? If \code{FALSE}, bandwidth is cross-validated using the
procedures from package \code{locpol}. The choice \code{TRUE} is compatible
with the bandwidth selection procedure in \code{fllr}.}

\item{fullgrid}{logical; should the full grid [0,1] of k-NN be used
directly for cross-validation? If \code{FALSE}, the grid is built iteratively
from small to larger \code{k}. By default, \code{fullgrid=FALSE}.}
}
\value{
A list with the following components:
\itemize{
\item \code{Predicted.responses} A vector of predicted responses, each
element of the vector corresponding to a row of \code{PRED}.
\item \code{PREDICTED.DERIV} A matrix of predicted (Riesz representations
of) functional derivatives, each row corresponding to a row of \code{PRED}.
\item \code{PREDICTED.DERIV.LIN} Only if \code{knn.ind=TRUE}. A matrix of
predicted (Riesz representations of) functional derivatives, each row
corresponding to a row of \code{PRED}. Instead of in \code{PREDICTED.DERIV}
where the local quadratic estimator is employed, here the local linear
estimator is used. Only for compatibility purposes.
\item \code{dim} Dimension of the eigenbasis used for both the regression
operator and the functional derivative. Chosen as a given percentage of total
sample variance explained by the first eigenfunctions of the empirical
covariance operator (if \code{method}=\code{FVE}), or chosen using
the conditional pseudo-Gaussian BIC (if \code{method}=\code{BIC}).
}
}
\description{
Estimates the scalar-on-function model using the additive modeling approach
from Mueller and Yao (2008) for the regression operator, and Mueller and Yao
(2010) for the functional derivatives.
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses

res = FAM(Responses, LEARN, PRED)

plot(res$Predicted.responses~dat$oracle.Responses)

}
\references{
Mueller, H.-G., and Yao, F. (2008). Functional additive models.
\emph{Journal of the American Statistical Association} 103, 1534--1544.

Mueller, H.-G., and Yao, F. (2010). Additive modelling of
functional gradients. \emph{Biometrika} 97, 791--805.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
