% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fllr.R
\name{fllr}
\alias{fllr}
\title{Functional local linear regression}
\usage{
fllr(
  Responses,
  LEARN,
  PRED,
  nboot = 100,
  kNN.grid.length = 30,
  percent = 0.95,
  boot.seed = NULL,
  BASIS = NULL,
  method = "cv",
  derivative = TRUE,
  J.est = TRUE,
  fullgrid = NULL
)
}
\arguments{
\item{Responses}{vector of scalar responses}

\item{LEARN}{matrix of regressors, functional values evaluated at common
domain, one function per row}

\item{PRED}{matrix of functions at which new values are predicted, one
function per row}

\item{nboot}{number of wild bootstrap replicates used for finding the
bandwidth for functional derivatives}

\item{kNN.grid.length}{grid length for the bandwidth search}

\item{percent}{if \code{BASIS} is not given, percentage of explained sample
variance that is taken into account when building the basis expansion in
terms of eigenfunctions of the empirical covariance operator}

\item{boot.seed}{random seed for the bootstrap procedure}

\item{BASIS}{if basis functions are not estimated using the empirical
covariance operator, matrix of basis functions. One function per row.}

\item{method}{bandwidth selection method for the regression operator.
Possible values are \code{cv} for leave-one-out cross-validation, and
\code{aicc} for the improved Akaike information criterion}

\item{derivative}{logical; should also the bandwidth for the functional
derivative be searched for? If set to \code{FALSE}, the bandwidth for the
regression operator is used also for the functional derivatives.}

\item{J.est}{logical; should it be cross-validated also with respect to
the dimension of the basis? If set to \code{FALSE}, the number of functions
in \code{BASIS} is taken to be the optimal dimension. Ignored if
\code{BASIS = NULL}.}

\item{fullgrid}{logical; should the full grid [0,1] of k-NN be used
directly for cross-validation? If \code{FALSE}, the grid is built iteratively
from small to larger \code{k}. By default, \code{fullgrid=TRUE} if the number
of functions in \code{LEARN} is at most 250.}
}
\value{
A list with the following components:
\itemize{
\item \code{Estimated.responses} A vector of responses estimated by the local
linear method that correspond to the functions in the rows of \code{LEARN}.
\item \code{ESTIMATED.DERIV} A matrix of the estimated (Riesz representations
of) functional derivatives of the regression operator, each row corresponding
to a row of \code{LEARN}.
\item \code{Predicted.responses} If matrices \code{LEARN} and \code{PRED}
differ, a vector of predicted responses, each element of the vector
corresponding to a row of \code{PRED}.
\item \code{PREDICTED.DERIV} If matrices \code{LEARN} and \code{PRED} differ,
a matrix of predicted (Riesz representations of) functional derivatives,
each row corresponding to a row of \code{PRED}.
\item \code{min.crit.reg} Minimum criterion value (mean squared error for
\code{method = cv} or AICc for \code{method = aicc}) obtained in the search
for the optimal bandwidth for the regression operator.
\item \code{knn.opt} List of optimal bandwidths for the regression operator
\code{reg} and the functional derivative \code{deriv}.
\item \code{dim.opt} Optimal basis dimension used for both the regression
operator and the functional derivative.
}
}
\description{
Functional local linear regression (fllr) with automatic bandwidth selection
including data-driven basis (functional predictor's FPCA) and a data driven
selection of the dimension of the approximating subspace.
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses

res = fllr(Responses, LEARN, PRED)

plot(res$Predicted.responses~dat$oracle.Responses)

}
\references{
Ferraty, F., and Nagy, S. (2020).
Scalar-on-function local linear regression and beyond.
\emph{Under review}.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
