% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Data simulation}
\usage{
generate(
  nlearn = 100,
  ntest = 500,
  Jmodel = 4,
  nsr = 0.05,
  aa = 1,
  rho = NULL,
  model = "M3"
)
}
\arguments{
\item{nlearn}{number of learning functions}

\item{ntest}{number of testing functions}

\item{Jmodel}{effective dimension of the regressors, number of basis
functions on which the responses truly depend}

\item{nsr}{noise-to-signal ration, a number between 0 and 1}

\item{aa}{coefficient a in model \code{M3}, determining the linear
combination of the linear and the non-linear part of the regression}

\item{rho}{parameter that drives the noise in the confounding part of
the regressors in models \code{M2} and \code{M3}, i.e. the part of the
regressors that corresponds to basis functions that do not contribute to the
modelling of the response}

\item{model}{indicator of the model used, accepted are \code{"M1"},
\code{"M2"} or \code{"M3"}}
}
\value{
A list with the following components:
\itemize{
\item \code{LEARN} Matrix of dimension nlearn-times-101, one learning
function per row.
\item \code{PRED}  Matrix of new regressors of dimension npred-times-101, one
function per row.
\item \code{Responses} Vector of responses that correspond to the rows of
\code{LEARN}.
\item \code{oracle.Responses} True (unknown) responses generated by the model
that correspond to the rows of \code{PRED}.
\item \code{oracle.DERIV} True (unknown) functional derivatives generated
by the model that correspond to the rows of \code{PRED}.
\item \code{fDERIV = DERIV} The complete matrix of functional derivatives of
dimension (nlearn+npred)-times-101, first nlearn functions correspond to the
rows of \code{LEARN}, the remaining npred function to the rows of
\code{PRED}.
\item \code{fRegression} The complete vector of noiseless conditional
expectations in the model of length (nlearn+npred), first nlearn values
correspond to the rows of \code{LEARN}, the remaining npred values to the
rows of \code{PRED}.
\item \code{DERIV} True functional derivatives generated by the model.
Matrix of dimension nlearn-times-101, one row per a row of \code{LEARN}.
\item \code{BASIS} Matrix of the basis vectors into which the functional
regressors were decomposed.
}
}
\description{
Simulates a random sample from models (M1)-(M3) from Ferraty and Nagy (2020).
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses

res = fllr(Responses, LEARN, PRED)

plot(res$Predicted.responses~dat$oracle.Responses)

}
\references{
Ferraty, F., and Nagy, S. (2020).
Scalar-on-function local linear regression and beyond.
\emph{Under review}.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
