% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPR.R
\name{GPR}
\alias{GPR}
\title{Gaussian process regression}
\usage{
GPR(Responses, LEARN, PRED)
}
\arguments{
\item{Responses}{vector of scalar responses}

\item{LEARN}{matrix of regressors, functional values evaluated at common
domain, one function per row}

\item{PRED}{matrix of functions at which new values are predicted, one
function per row}
}
\value{
\code{Predicted.responses} A vector of predicted responses, each
element of the vector corresponding to a row of \code{PRED}.
}
\description{
Estimates the scalar-on-function linear model using a Gaussian process method
for a single functional predictor, and a scalar response.
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses

res = GPR(Responses, LEARN, PRED)

plot(res$Predicted.responses~dat$oracle.Responses)

}
\references{
Wang, B., and Xu, A. (2019).
Gaussian process methods for nonparametric functional regression with mixed
predictors.	\emph{Computational Statistics & Data Analysis}, 131, 80--90.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
