% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelSmoother.R
\name{KernelSmoother}
\alias{KernelSmoother}
\title{Local constant functional estimation}
\usage{
KernelSmoother(
  Responses,
  DIST,
  DIST.PRED,
  kNN.grid.length = 30,
  fullgrid = FALSE
)
}
\arguments{
\item{Responses}{vector of scalar responses}

\item{DIST}{a square symmetric matrix of all distances of regressor
functions from each other, see \code{\link{L2metric}}.}

\item{DIST.PRED}{an \code{m}-times-\code{n} matrix of distances of all
functions at which we intend to predict (\code{m} functions) from the
regressor functions (\code{n} functions).}

\item{kNN.grid.length}{grid length for the bandwidth search.}

\item{fullgrid}{logical; should the full grid [0,1] of k-NN be used
directly for cross-validation? If \code{FALSE}, the grid is built iteratively
from small to larger \code{k}. By default, \code{fullgrid=FALSE}.}
}
\value{
A list with the following components:
\itemize{
\item \code{Predicted.responses} A vector of predicted responses, each
element of the vector corresponding to a function that corresponds to a row
in \code{DIST}.
\item \code{mse} Minimum mean squared error obtained in the search for the
optimal bandwidth for the regression operator.
\item \code{knn.opt} The optimal bandwidths for the regression operator
obtained by leave-one-out cross-validation.
}
}
\description{
Estimates the scalar-on-function model using the local constant
(Nadaraya-Watson's estimator) for the regression operator and functional
data. Does not directly provide an estimator of functional derivatives.
Bandwidth is chosen via leave-one-out cross-validation.
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses

DIST = L2metric(LEARN,LEARN)
DIST.PRED = L2metric(PRED,LEARN)
res = KernelSmoother(Responses, DIST, DIST.PRED)

plot(res$Predicted.responses~dat$oracle.Responses)

}
\references{
Ferraty, F., and Vieu, P. (2006). Nonparametric functional data
analysis. \emph{Springer Series in Statistics. Springer, New York}.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
