% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelSmoother.R
\name{ksm}
\alias{ksm}
\title{Local constant smoother with a fixed bandwidth}
\usage{
ksm(
  D,
  Y,
  h = NULL,
  kNN = NULL,
  kernel = c("uniform", "triangular", "Epanechnikov", "biweight", "triweight",
    "Gaussian"),
  quantile.type = 7
)
}
\arguments{
\item{D}{a (possibly non-symmetric) matrix of distances of the
learning sample of functional data from the predictor functions, one row per
a learning sample function}

\item{Y}{vector of scalar responses}

\item{h}{a single bandwidth. Alternatively, if \code{kNN} is
specified, local bandwidths are built automatically using the nearest
neighbors.}

\item{kNN}{number of nearest neighbors (as a fraction of total number
of neighbors, i.e. a number between 0 and 1) to be considered for local
bandwidth selection. If \code{kNN} is provided, vector \code{h} is build
automatically by evaluating sample quantiles of the distances in rows of the
matrix of distances \code{D}.}

\item{kernel}{kernel function. Possible choices are \code{"uniform"},
\code{"triangular"}, \code{"Epanechnikov"}, \code{"biweight"},
\code{"triweight"}, and \code{"Gaussian"}.}

\item{quantile.type}{if \code{kNN} is provided, parameter that enters
function \code{\link{quantile}} that specifies the type of the sample
quantile taken.}
}
\value{
A vector of predicted responses whose length is the number of columns
of \code{D}.
}
\description{
The local constant smoother (Nadaraya-Watson's estimator)
in the scalar-on-function linear model with a given bandwidth.
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses

ksm(L2metric(LEARN, PRED), Responses, kNN = .2, kernel = "Gauss")

}
\references{
Ferraty, F., and Vieu, P. (2006). Nonparametric functional data
analysis. \emph{Springer Series in Statistics. Springer, New York}.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
