% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelSmoother.R
\name{ksm_cv}
\alias{ksm_cv}
\title{Cross-validation for the local constant smoother}
\usage{
ksm_cv(
  D,
  Y,
  H = NULL,
  kNN = NULL,
  nCV = NULL,
  kernel = c("uniform", "triangular", "Epanechnikov", "biweight", "triweight",
    "Gaussian"),
  quantile.type = 7
)
}
\arguments{
\item{D}{symmetric square matrix of distances of the functional data
from each other, one row per function}

\item{Y}{vector of scalar responses}

\item{H}{bandwidths to be considered. Could be either a vector of
bandwidths that will be applied to each function, or a matrix with individual
bandwidths applied to learning function, one row per function. Alternatively,
if \code{kNN} is specified, the bandwidth matrix is built automatically using
the local nearest neighbors bandwidths.}

\item{kNN}{sequence of nearest neighbors (as a fraction of total number
of neighbors, i.e. a number between 0 and 1) to be considered for local
bandwidth selection. If \code{kNN} is provided, matrix \code{H} is build
automatically by evaluating sample quantiles of the distances in rows of
the matrix of distances \code{D}.}

\item{nCV}{if provided, only the first \code{nCV} leave-one-out
cross-validated trials is considered in the resulting mean squared error.
By default, all functions are included in cross-validation.}

\item{kernel}{kernel function. Possible choices are \code{"uniform"},
\code{"triangular"}, \code{"Epanechnikov"}, \code{"biweight"},
\code{"triweight"}, and \code{"Gaussian"}.}

\item{quantile.type}{if \code{kNN} is provided, parameter that enters
function \code{\link{quantile}} that specifies the type of the sample
quantile taken.}
}
\value{
A list with the following components:
\itemize{
\item \code{CV} A matrix that corresponds to \code{H} that contains all the
resulting mean squared errors of the cross-validation procedure.
\item \code{kernel} Kernel used.
}
}
\description{
For the local constant smoother (Nadaraya-Watson's estimator)
in the scalar-on-function linear model, given a sequence of bandwidths,
returns a sequence of (leave-one-out) cross-validated mean squared errors
for bandwidth selection.
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses

ksm_cv(L2metric(LEARN, LEARN), Responses, kNN = seq(.1,.9,length=11),
kernel = "Gauss")

}
\references{
Ferraty, F., and Vieu, P. (2006). Nonparametric functional data
analysis. \emph{Springer Series in Statistics. Springer, New York}.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
