% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{L2metric}
\alias{L2metric}
\title{\code{L2}-metric}
\usage{
L2metric(A, B, correct = TRUE)
}
\arguments{
\item{A}{A matrix of functional values, one function per row, \code{n} rows.}

\item{B}{A matrix of functional values, one function per row, \code{m} rows.}

\item{correct}{Logical; should the boundary points be computed only by a half of their
value due to boundary effects?}
}
\value{
A matrix of size \code{n}-times-\code{m} that contains approximations to the
\code{L2}-distances between function that correspond to the rows of \code{A} and \code{B}.
The functions must be evaluated at an equidistant grid. It may be necessary to multiply
by a constant that corresponds to the grid size.
}
\description{
Fast computation of the \code{L2}-metric between two sets of functional data.
}
\examples{
gridsize = 101                           # grid size
J = 4                                    # number of basis elements
BASIS = phif(Jmodel, Grid)
L2metric(BASIS, BASIS)
L2metric(BASIS, BASIS, correct = FALSE)
sqrt(crossprod(BASIS[1,]-BASIS[2,]))

}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
