% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModel.R
\name{LinearModel}
\alias{LinearModel}
\title{Functional linear model}
\usage{
LinearModel(Responses, LEARN, PRED, percent = 0.95, BASIS = NULL)
}
\arguments{
\item{Responses}{vector of scalar responses}

\item{LEARN}{matrix of regressors, functional values evaluated at common
domain, one function per row}

\item{PRED}{matrix of functions at which new values are predicted, one
function per row}

\item{percent}{if \code{BASIS} is not given, percentage of explained sample
variance that is taken into account when building the basis expansion in
terms of eigenfunctions of the empirical covariance operator}

\item{BASIS}{if basis functions are not estimated using the empirical
covariance operator, matrix of basis functions. One function per row.}
}
\value{
A list with the following components:
\itemize{
\item \code{Predicted.responses} A vector of predicted responses, each
element of the vector corresponding to a row of \code{PRED}.
\item \code{PREDICTED.DERIV} A matrix of predicted (Riesz representations of)
functional derivatives, each row corresponding to a row of \code{PRED}.
\item \code{mse} Minimum mean squared error obtained in the search
leave-one-out cross-validation for the optimal dimension of the basis for
the regression operator.
\item \code{dim.opt} Optimal basis dimension used for both the regression
operator and the functional derivative. If \code{BASIS} is provided,
\code{dim.opt} is chosen as the numer of rows of \code{BASIS}.
}
}
\description{
Estimates the standard scalar-on-function linear model, where all functional
data are projected into a subspace given by basis functions.
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses

res = LinearModel(Responses, LEARN, PRED)

plot(res$Predicted.responses~dat$oracle.Responses)

}
\references{
Reiss, P. T., and Ogden, R. T. (2007). Functional principal
component regression and functional partial least squares. \emph{Journal of
the American Statistical Association}. 102, 984--996.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
