% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{llsm}
\alias{llsm}
\title{Functional local linear smoother with a fixed bandwidth}
\usage{
llsm(
  C,
  Cnew,
  D,
  Y,
  h = NULL,
  kNN = NULL,
  kernel = c("uniform", "triangular", "Epanechnikov", "biweight", "triweight",
    "Gaussian"),
  quantile.type = 7
)
}
\arguments{
\item{C}{a matrix of the basis coefficients of the the learning sample
functions (n functions), one row per function}

\item{Cnew}{a matrix of the basis coefficients of the predictor functions
(m functions), one row per predictor}

\item{D}{a matrix of distances of the predictor functions from the
learning sample functions (m-times-n matrix), one row per a predictor}

\item{Y}{vector of scalar responses}

\item{h}{either a single bandwidth, or a vector of m bandwidths to be used
for the predictor functions. Alternatively, if \code{kNN} is specified,
local bandwidths are built automatically using the nearest neighbors.}

\item{kNN}{number of nearest neighbors (as a fraction of total number of neighbors, i.e.
a number between 0 and 1) to be considered for local bandwidth
selection. If \code{kNN} is provided, vector \code{h} is build automatically by
evaluating sample quantiles of the distances in rows of the matrix of distances \code{D}.}

\item{kernel}{kernel function. Possible choices are \code{"uniform"}, \code{"triangular"},
\code{"Epanechnikov"}, \code{"biweight"}, \code{"triweight"}, and \code{"Gaussian"}.}

\item{quantile.type}{if \code{kNN} is provided, parameter that enters function \code{\link{quantile}}
that specifies the type of the sample quantile taken}
}
\value{
A matrix where each row corresponds to a single predictor.
The columns are the estimated regression coefficients in the given basis expansion.
The first column is the intercept - an estimate of the regression operator. Remaining
columns correspond to the basis expansion of the estimator of the functional derivative
of the regression operator.
}
\description{
The functional local linear smoother
in the scalar-on-function linear model with a given bandwidth.
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses
BASIS = dat$BASIS

C = project(LEARN,BASIS)
Cnew = project(PRED,BASIS)
D = L2metric(PRED,LEARN)
res = llsm(C, Cnew, D, Responses, kNN = .05, kernel = "Gauss")

plot(res[,1]~Regression[-(1:nlearn)])
}
\references{
Ferraty, F., and Nagy, S. (2021).
Scalar-on-function local linear regression and beyond.
\emph{Biometrika}, to appear.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
