% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{llsm_cv}
\alias{llsm_cv}
\title{Cross-validation for the functional local linear smoother}
\usage{
llsm_cv(
  C,
  D,
  Y,
  H = NULL,
  kNN = NULL,
  nCV = NULL,
  kernel = c("uniform", "triangular", "Epanechnikov", "biweight", "triweight",
    "Gaussian"),
  quantile.type = 7
)
}
\arguments{
\item{C}{a matrix of the basis coefficients of the the learning sample
functions (n functions), one row per function}

\item{D}{a symmetric square matrix of distances of the learning sample
functions from themselves (n-times-n matrix), one row per a function}

\item{Y}{vector of scalar responses}

\item{H}{bandwidths to be considered. Could be either a vector of bandwidths that
will be applied to each function, or a matrix with individual bandwidths applied to
learning functions, one row per function. Alternatively, if \code{kNN} is specified,
the bandwidth matrix is built automatically using the local nearest neighbors bandwidths.}

\item{kNN}{sequence of nearest neighbors (as a fraction of total number of neighbors, i.e.
a number between 0 and 1) to be considered for local bandwidth
selection. If \code{kNN} is provided, matrix \code{H} is build automatically by
evaluating sample quantiles of the distances in rows of the matrix of distances \code{D}.}

\item{nCV}{if provided, only the first \code{nCV} leave-one-out cross-validated
trials is considered in the resulting mean squared error. By default, all functions are
included in cross-validation.}

\item{kernel}{kernel function. Possible choices are \code{"uniform"}, \code{"triangular"},
\code{"Epanechnikov"}, \code{"biweight"}, \code{"triweight"}, and \code{"Gaussian"}.}

\item{quantile.type}{if \code{kNN} is provided, parameter that enters function \code{\link{quantile}}
that specifies the type of the sample quantile taken}
}
\value{
A list with the following components:
\itemize{
\item \code{CV} A matrix that corresponds to \code{H} that contains all the resulting
mean squared errors of the cross-validation procedure.
\item \code{CVB} A matrix that corresponds to \code{H} that contains all the resulting
variance estimates (mean squared errors - bias squared) of the cross-validation procedure.
\item \code{kernel} Kernel used.
}
}
\description{
For the functional local linear smoother
in the scalar-on-function linear model, given a sequence of bandwidths,
returns a sequence of (leave-one-out) cross-validated mean squared errors
for bandwidth selection.
}
\examples{
dat = generate(100,200,4)
LEARN = dat$LEARN
PRED = dat$PRED
Responses = dat$Responses
BASIS = dat$BASIS

C = project(LEARN,BASIS)
D = L2metric(LEARN,LEARN)
llsm_cv(C, D, Responses, kNN = seq(.1,.9,length=11), kernel = "Gauss")

}
\references{
Ferraty, F., and Nagy, S. (2020).
Scalar-on-function local linear regression and beyond.
\emph{Under review}.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
