% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{oracle.bw}
\alias{oracle.bw}
\title{Oracle bandwidth value for functional derivative estimation}
\usage{
oracle.bw(
  Responses,
  LEARN,
  ORACLE.DERIV,
  BASIS,
  kNN.grid.length = 30,
  fullgrid = FALSE
)
}
\arguments{
\item{Responses}{vector of scalar responses}

\item{LEARN}{matrix of regressors, functional values evaluated at common domain,
one function per row}

\item{ORACLE.DERIV}{matrix of the true functional derivatives, each row corresponding
to a row of \code{LEARN}.}

\item{BASIS}{basis functions into which the random functions are decomposed. One
function per row. The true dimension of this space is assumed to be the number of rows
of this matrix.}

\item{kNN.grid.length}{grid length for the bandwidth search}
}
\value{
A single bandwidth in terms of the number of nearest neighbors that
corresponds to \code{knn.opt$deriv} from \code{\link{fllr}}. Works for
\code{method = cv} only.
}
\description{
Given the true (Riesz representatives of) functional derivatives in a scalar-on-function
regression problem, this function evaluates the oracle bandwidth choice for the estimation
of the functional derivatives in a given functional basis.
}
\references{
Ferraty, F., and Nagy, S. (2021).
Scalar-on-function local linear regression and beyond.
\emph{Biometrika}, to appear.
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
