% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{phif}
\alias{phif}
\title{Fourier basis functions}
\usage{
phif(J, t)
}
\arguments{
\item{J}{Number of basis functions.}

\item{t}{Vector of the common points in the domain where the basis functions are evaluated.}
}
\value{
A matrix with \code{J} rows. In each row there are functional values of a basis
function that correspond to the points in the domain from \code{t}.
}
\description{
Creates a matrix of the first few functions from the Fourier basis.
}
\examples{
gridsize = 101                           # grid size
J = 4                                    # number of basis elements
BASIS = phif(Jmodel, Grid)               # first J Fourier basis elements
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
