% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{project}
\alias{project}
\title{Basis projection}
\usage{
project(X, basis)
}
\arguments{
\item{X}{a vector of a single function, or a matrix with one function per row}

\item{basis}{a matrix of basis functions, one function per row}
}
\value{
A matrix of coefficients of the functions in \code{X} with respect to the
basis in \code{basis}.
}
\description{
Project a set of functions onto a basis.
}
\examples{
gridsize = 101                           # grid size
Grid = seq(0, 1, length = gridsize)      # grid of measurements
n = 150
Jmodel = 4                               # number of basis elements
BASIS = phif(Jmodel, Grid)               # first Jmodel Fourier basis elements
UNIF = matrix(runif(sample.size * Jmodel, min = -1, max = 1), nrow = sample.size, ncol = Jmodel)

PREDICTORS = UNIF \%*\% BASIS[1:Jmodel, ]
PREDICTORS2 = expand(UNIF,BASIS[1:Jmodel, ])
all.equal(PREDICTORS,PREDICTORS2)        # expansion of coefficients into functions

COEF = project(PREDICTORS,BASIS)         # projection back to basis functions
max(abs(COEF-UNIF))
}
\author{
Stanislav Nagy, \email{nagy at karlin.mff.cuni.cz}
}
